restricciones de acceso, nav y boton inicio sesion interctivo

+++ src/app/componentes/listar-usuarios/listar-usuarios.component.ts
@@ -1,5 +1,6 @@
 import { Component } from '@angular/core';
 import { Router } from '@angular/router';
+import { TokenService } from 'src/app/servicios/token/token.service';
 import { EstadoUsuariosService } from 'src/app/servicios/usuario/estado-usuarios.service';
 import { RolService } from 'src/app/servicios/usuario/roles.service';
 import { UsuarioService } from 'src/app/servicios/usuario/usuario.service';
@@ -18,6 +19,7 @@ export class ListarUsuariosComponent {
   constructor(
     private usuarioService: UsuarioService,
     private router: Router,
+    private tokenService: TokenService,
     private rolesService: RolService,
     private estadoUsuarios: EstadoUsuariosService
   ) { }
@@ -65,19 +67,23 @@ export class ListarUsuariosComponent {
     const rol = this.roles.find((rol) => rol.id_rol === idRol);
     return rol ? rol.rol : '';
   }
-  
+
   obtenerNombreEstado(idEstado: number): string {
     const estado = this.estados.find((estado) => estado.id_estado_usuario === idEstado);
     return estado ? estado.estado : '';
   }
 
   eliminarUsuario(id_usuario: number) {
-    console.log(`Eliminar usuario con ID ${id_usuario}`);
-    if (confirm('¿Está seguro de que desea eliminar este usuario?')) {
-      this.usuarioService.eliminarUsuario(id_usuario).subscribe(() => {
-        console.log(`Usuario con ID ${id_usuario} eliminado`);
-        this.obtenerUsuarios();
-      });
+    if (this.tokenService.verSesionUsuario(id_usuario)) {
+      console.log(`Eliminar usuario con ID ${id_usuario}`);
+      if (confirm('¿Está seguro de que desea eliminar este usuario?')) {
+        this.usuarioService.eliminarUsuario(id_usuario).subscribe(() => {
+          console.log(`Usuario con ID ${id_usuario} eliminado`);
+          this.obtenerUsuarios();
+        });
+      }
+    }else{
+      console.log(`No puedes borrar un usuaio que esta logueado usuario`);
     }
   }
 }
